+++ examples/python/using_qiskit_terra_level_2.py
@@ -19,11 +19,10 @@ as making a pass manager and telling it to start with CXCancellation.
 from qiskit import IBMQ, qobj_to_circuits
 
 try:
-    import Qconfig
-    IBMQ.enable_account(Qconfig.APItoken, Qconfig.config['url'])
+    IBMQ.load_accounts()
 except:
     print("""WARNING: There's no connection with the API for remote backends.
-             Have you initialized a Qconfig.py file with your personal token?
+             Have you initialized a file with your personal token?
              For now, there's only access to local simulator backends...""")
 
 
@@ -45,14 +44,14 @@ print(circ.qasm())
 
 # 1. standard compile -- standard qiskit passes, when no PassManager given
 from qiskit import transpiler
-qobj_standard = transpiler.compile(circ, backend_device)
-[compiled_standard] = qobj_to_circuits(qobj_standard)
+dags = transpiler.compile(circ, backend_device)
+[compiled_standard] = dags
 print(compiled_standard.qasm())
 
 # 2. custom compile -- customize PassManager to run specific circuit transformations
 from qiskit.transpiler.passes import CXCancellation
 pm = transpiler.PassManager()
 pm.add_passes(CXCancellation())
-qobj_custom = transpiler.compile(circ, backend_device, pass_manager=pm)
-[compiled_custom] = qobj_to_circuits(qobj_custom)
+dags = transpiler.compile(circ, backend_device, pass_manager=pm)
+[compiled_custom] = dags
 print(compiled_custom.qasm())
