+++ qiskit/wrapper/__init__.py
@@ -7,7 +7,7 @@
 
 # pylint: disable=redefined-builtin
 
-"""Helper module for simplified QISKit usage.
+"""Helper module for simplified Qiskit usage.
 
 The functions in this module provide convenience helpers for accessing commonly
 used features of the SDK in a simplified way. They support a small subset of
@@ -15,4 +15,4 @@ scenarios and flows: for more advanced usage, it is encouraged to instead
 refer to the documentation of each component and use them separately.
 """
 
-from ._wrapper import (compile, execute, load_qasm_string, load_qasm_file, qobj_to_circuits)
+from ._wrapper import (load_qasm_string, load_qasm_file, qobj_to_circuits)
