+++ test/python/aer/test_extensions_simulator.py
@@ -14,7 +14,7 @@ import qiskit
 import qiskit.extensions.simulator
 from qiskit import Aer
 from qiskit.tools.qi.qi import state_fidelity
-from qiskit.wrapper import execute
+from qiskit import execute
 from ..common import QiskitTestCase, requires_cpp_simulator
