+++ test/python/test_initializer.py
@@ -18,7 +18,7 @@ from qiskit import QISKitError
 from qiskit import QuantumCircuit
 from qiskit import QuantumRegister
 from qiskit import ClassicalRegister
-from qiskit import wrapper, Aer
+from qiskit import execute, Aer, compile
 from qiskit.tools.qi.qi import state_fidelity
 from .common import QiskitTestCase
 
@@ -33,7 +33,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(2, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0], qr[1]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -46,7 +46,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(2, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0], qr[1]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -59,7 +59,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(2, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0], qr[1]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -72,7 +72,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(3, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0], qr[1], qr[2]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -85,7 +85,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(1, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -106,7 +106,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(3, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0], qr[1], qr[2]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -135,7 +135,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(4, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0], qr[1], qr[2], qr[3]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -173,7 +173,7 @@ class TestInitialize(QiskitTestCase):
         # statevector simulator does not support reset
         shots = 2000
         threshold = 0.04 * shots
-        job = wrapper.execute(qc, Aer.get_backend('qasm_simulator'), shots=shots)
+        job = execute(qc, Aer.get_backend('qasm_simulator'), shots=shots)
         result = job.result()
         counts = result.get_counts()
         target = {'00': shots / 4, '01': shots / 4, '10': shots / 4, '11': shots / 4}
@@ -200,7 +200,7 @@ class TestInitialize(QiskitTestCase):
         qr = QuantumRegister(4, "qr")
         qc = QuantumCircuit(qr)
         qc.initialize(desired_vector, [qr[0], qr[1], qr[2], qr[3]])
-        job = wrapper.execute(qc, Aer.get_backend('statevector_simulator'))
+        job = execute(qc, Aer.get_backend('statevector_simulator'))
         result = job.result()
         statevector = result.get_statevector()
         fidelity = state_fidelity(statevector, desired_vector)
@@ -218,7 +218,7 @@ class TestInitialize(QiskitTestCase):
         qc2 = QuantumCircuit(qr, cr)
         qc2.initialize([1.0 / math.sqrt(2), -1.0 / math.sqrt(2)], [qr[0]])
 
-        job = wrapper.execute(qc1+qc2, Aer.get_backend('statevector_simulator'))
+        job = execute(qc1+qc2, Aer.get_backend('statevector_simulator'))
         result = job.result()
         quantum_state = result.get_statevector()
         fidelity = state_fidelity(quantum_state, desired_vector)
